refactor(dashboard): extract org_onboard + org_seal sub-routers (F-B-202 PR-6/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-6 of 10. base=main.

Extracts 7 routes (3 org onboard + 4 org seal/unseal-reauth):

  app/dashboard/org_onboard_routes.py (new, 235 LOC, 3 routes):
    POST /dashboard/orgs/onboard/generate-ca   mint self-signed P-256 CA
    GET  /dashboard/orgs/onboard               onboarding form
    POST /dashboard/orgs/onboard               register org + attach CA

  app/dashboard/org_seal_routes.py (new, 172 LOC, 4 routes):
    GET  /dashboard/orgs/{id}/unseal-reauth    password prompt form
    POST /dashboard/orgs/{id}/unseal-reauth    verify pwd + stamp scope
    POST /dashboard/orgs/{id}/seal             force-seal (no re-auth)
    POST /dashboard/orgs/{id}/unseal           unseal (sealed-gate)

  app/dashboard/router.py (2372 → 2129 LOC, -243):
    - 7 routes removed; breadcrumb comments at the old positions.
    - Two new include_router calls.

Behavioural delta: zero. Audit F-B-2 sealed-org guard semantics
preserved on /unseal (still calls _require_sealed_reauth before
flipping the sealed flag off).

Pattern note on _DISPLAY_NAME_MAX: the constant is duplicated into
org_onboard_routes.py (kept at 256) so the sub-router stays self-
contained. Future micro-refactor can consolidate it via
app/dashboard/_helpers.py if more sub-routers consume it.

Tests:
  - pytest tests/test_dashboard*.py -q — 202/203 (1 pre-existing deselect)
  - pytest tests/ <full suite> -q — 4216 passed, 9 skipped, 32 xfailed
    (1 xdist flake pre-existing, passes isolated)

Conflict warning: this PR starts from PR-4-merged (PR-5 #848 still open).
Rebase on main post-#848 expected to be clean — sections 18+20 have
zero line overlap with PR-5 sections 9+10+19+21.

Router LOC progress:
  Pre-Sprint 2 : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC
  Post-PR-2 #841: 60 routes, 2614 LOC
  Post-PR-3 #842: 57 routes, 2528 LOC
  Post-PR-4 #847: 50 routes, 2372 LOC
  Post-PR-5 #848 (still open): 39 routes, 2071 LOC
  Post-PR-6 (this, before #848 rebase): 43 routes, 2129 LOC
  Post-PR-5 + PR-6 merged: 32 routes, ~1828 LOC
  Target post-PR-10: ~3 (aggregator), ~100 LOC

Next PR-7 (F-B-202 modularization): extract agents_lifecycle_routes.py
+ agents_credentials_routes.py (sections 22+23+24, ~894 LOC, 12 routes
— biggest extract).
